# perf: Faster ImmutableList.Add