Benchmark:

```csharp
    private static readonly object[] s_arr = new object[] { new(), new(), new(), new(), new(), new(), new() };
    private static readonly object s_obj = new();
    [Benchmark(Baseline = true)]
    public void WithoutSpan()
    {
        var arr = new object[s_arr.Length + 1];
        Array.Copy(s_arr, arr, s_arr.Length);
        arr[s_arr.Length] = s_obj;
    }

    [Benchmark]
    public void WithSpan()
    {
        var arr = new object[s_arr.Length + 1];
        Span<object> arrSpan = MemoryMarshal.CreateSpan(ref MemoryMarshal.GetArrayDataReference(arr), arr.Length);
        Span<object> span = MemoryMarshal.CreateSpan(ref MemoryMarshal.GetArrayDataReference(s_arr), s_arr.Length);
        span.CopyTo(arrSpan);
        arrSpan[s_arr.Length] = s_obj;
    }
```

|      Method |     Mean |    Error |   StdDev |   Median | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|---------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|
| WithoutSpan | 14.83 ns | 0.344 ns | 0.970 ns | 14.35 ns |  1.00 |    0.00 | 0.0421 |     - |     - |      88 B |
|    WithSpan | 11.40 ns | 0.160 ns | 0.125 ns | 11.37 ns |  0.70 |    0.03 | 0.0421 |     - |     - |      88 B |